backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
